+++ src/fan_temperature_control.h
@@ -19,8 +19,7 @@ struct FanTemperatureControl {
 typedef struct FanTemperatureControl FanTemperatureControl;
 declare_array_of(FanTemperatureControl);
 
-Error* FanTemperatureControl_SetDefaults(FanTemperatureControl*, int);
-Error* FanTemperatureControl_SetByConfig(array_of(FanTemperatureControl)*, array_of(FanTemperatureSourceConfig)*, ModelConfig*);
+Error* FanTemperatureControl_Init(array_of(FanTemperatureControl)*, ServiceConfig*, ModelConfig*);
 Error* FanTemperatureControl_UpdateFanTemperature(FanTemperatureControl*);
 void   FanTemperatureControl_Log(array_of(FanTemperatureControl)*, ModelConfig*);